backend/requirements.txt
flask==3.1.*
flask-cors==5.*
flask-migrate==4.*
sqlalchemy==2.0.*
psycopg2-binary==2.9.*
redis==5.*
apscheduler==3.10.*
marshmallow==3.*
marshmallow-sqlalchemy==1.*
anthropic==0.40.*
python-dotenv==1.*
httpx==0.27.*
itsdangerous==2.*
gunicorn==22.*
structlog==24.*
